+++ fmpz/combit.c
@@ -31,7 +31,7 @@ void fmpz_combit(fmpz_t f, ulong i)
     {
         if (i < FLINT_BITS - 2)
         {
-            *f ^= (1L << i);
+            *f ^= (WORD(1) << i);
         }
         else  /* i >= FLINT_BITS */
         {
