+++ fmpz/init_set_readonly.c
@@ -43,7 +43,7 @@ void fmpz_init_set_readonly(fmpz_t f, const mpz_t z)
     }
     else
     {
-        *f = 0L;
+        *f = WORD(0);
     }
 }
