+++ fmpz_poly/taylor_shift_horner.c
@@ -33,19 +33,19 @@ _fmpz_poly_taylor_shift_horner(fmpz * poly, const fmpz_t c, slong n)
 {
     slong i, j;
 
-    if (*c == 1L)
+    if (*c == WORD(1))
     {
         for (i = n - 2; i >= 0; i--)
             for (j = i; j < n - 1; j++)
                 fmpz_add(poly + j, poly + j, poly + j + 1);
     }
-    else if (*c == -1L)
+    else if (*c == WORD(-1))
     {
         for (i = n - 2; i >= 0; i--)
             for (j = i; j < n - 1; j++)
                 fmpz_sub(poly + j, poly + j, poly + j + 1);
     }
-    else if (*c != 0L)
+    else if (*c != WORD(0))
     {
         for (i = n - 2; i >= 0; i--)
             for (j = i; j < n - 1; j++)
