+++ fmpz_poly_factor/realloc.c
@@ -27,7 +27,7 @@
 #include <stdlib.h>
 #include "flint.h"
 #include "fmpz.h"
-#include "fmpz_poly_factor.h"
+#include "fmpz_poly.h"
 
 void fmpz_poly_factor_realloc(fmpz_poly_factor_t fac, slong alloc)
 {
@@ -59,7 +59,7 @@ void fmpz_poly_factor_realloc(fmpz_poly_factor_t fac, slong alloc)
             for (i = fac->alloc; i < alloc; i++)
             {
                 fmpz_poly_init(fac->p + i);
-                fac->exp[i] = 0L;
+                fac->exp[i] = WORD(0);
             }
             fac->alloc = alloc;
         }
