+++ fmpz_vec.h
@@ -29,6 +29,7 @@
 
 #include <gmp.h>
 #include "fmpz.h"
+#include "flint.h"
 
 #ifdef __cplusplus
  extern "C" {
@@ -36,7 +37,7 @@
 
 #define FMPZ_VEC_NORM(vec, i)          \
 do {                                   \
-    while ((i) && vec[(i) - 1] == 0L)  \
+    while ((i) && vec[(i) - 1] == WORD(0))  \
         (i)--;                         \
 } while (0)
