+++ fmpz_vec/is_zero.c
@@ -33,7 +33,7 @@ _fmpz_vec_is_zero(const fmpz * vec, slong len)
 {
     slong i;
     for (i = 0; i < len; i++)
-        if (vec[i] != 0L)
+        if (vec[i] != WORD(0))
             return 0;
     return 1;
 }
